+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [2.x]
+- Add timezone information to datetime displayed in admin. (@saurabh)
 - Support for Django 2.0.x
+
 
 ## [1.11]